requirements.txt
torch==2.2.2
torchvision==0.17.2

fastapi==0.110.0
uvicorn==0.29.0
mlflow==2.11.0
pytest==8.1.1
pillow==10.2.0
tqdm==4.66.2

matplotlib==3.8.4
scikit-learn==1.4.2

seaborn
python-multipart==0.0.9
